[FLINK-25080][tests] Move ManuallyTriggeredScheduledExecutor to flink-core
